requirements.txt
crewai
crewai-tools
langchain_community
langchain_ollama
streamlit
litellm
python-dotenv
crewai[google-genai]